# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Meal Planner API",
    "description": "API for managing recipes, meal plans, and shopping lists",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/ingredients": {
      "get": {
        "summary": "List all ingredients",
        "tags": [
          "Ingredients"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "search",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "name": {
                        "type": "string"
                      },
                      "category": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "nullable": true,
                            "enum": [
                              null
                            ]
                          }
                        ]
                      },
                      "units": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "nullable": true,
                            "enum": [
                              null
                            ]
                          }
                        ]
                      },
                      "createdAt": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "nullable": true,
                            "enum": [
                              null
                            ]
                          }
                        ]
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "category",
                      "units",
                      "createdAt"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new ingredient",
        "tags": [
          "Ingredients"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "category": {
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "category": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "units": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "createdAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "category",
                    "units",
                    "createdAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/ingredients/{id}": {
      "get": {
        "summary": "Get an ingredient by ID",
        "tags": [
          "Ingredients"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "category": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "units": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "createdAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "category",
                    "units",
                    "createdAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update an ingredient",
        "tags": [
          "Ingredients"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "category": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "category": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "units": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "createdAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "category",
                    "units",
                    "createdAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an ingredient",
        "tags": [
          "Ingredients"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}